Move import declaration methods to new mixin.
